Add back "name" and add "ready?" to socket state
easier to have name in the state than generating it, update how we get
names from the Registry to determine how to send a message.

"ready?" indicates whether the socket conn has finished setting up with
the remote, which isn't done until the websocket struct is added to the
conn.