add Signer: a process holding private keys locally
the Signer module is a genserver that holds a private key, and provides a nip-46-like interface, namely `sign_event` and `get_pubkey`. It does not implement nip-44 or any DM-style NIPs, so this implementation is simply one step up from reading raw private keys from env and passing them to functions.
While this is "just a convenience" instead of passing private keys around, its implementation allows for any process that implements the Signer behaviour to act as a signer: this could be a remote process acting as a "bunker", without going through relays to sign.